+++ src/ViewBranches.ts
@@ -69,14 +69,14 @@ export class TreeViewBranches implements vscode.TreeDataProvider<Flow> {
                 .forEach(el => {
                     tree.push(new Flow(el,
                         el, 'git-branch', vscode.TreeItemCollapsibleState.None,
-                        this._isCurrent(el)
+                        this._isCurrent(el), 'branch'
                     ));
                 });
 
             tree.push(new Flow('release', 'Releases', 'tag', vscode.TreeItemCollapsibleState.Expanded, false, 'r'));
             tree.push(new Flow('feature', 'Features', 'test-view-icon', vscode.TreeItemCollapsibleState.Expanded, false, 'f'));
-            tree.push(new Flow('bugfix', 'BugFixes',  'callstack-view-session', vscode.TreeItemCollapsibleState.Expanded, false, 'b'));
-            tree.push(new Flow('hotfix', 'HotFixes',  'flame', vscode.TreeItemCollapsibleState.Expanded, false, 'h'));
+            tree.push(new Flow('bugfix', 'BugFixes', 'callstack-view-session', vscode.TreeItemCollapsibleState.Expanded, false, 'b'));
+            tree.push(new Flow('hotfix', 'HotFixes', 'flame', vscode.TreeItemCollapsibleState.Expanded, false, 'h'));
             return Promise.resolve(tree);
         }
 
@@ -109,6 +109,41 @@ export class TreeViewBranches implements vscode.TreeDataProvider<Flow> {
 
     fetchAllBranches() {
         this.util.exec("git fetch --all", true, s => {
+            this._onDidChangeTreeData.fire();
+        });
+    }
+
+    syncAll() {
+        vscode.window.withProgress({
+            location: vscode.ProgressLocation.Notification,
+            title: `Sync all root branches`,
+            cancellable: false
+        }, (progress, token) => new Promise<void>(resolve => {
+            setTimeout(() => {
+                this.listBranches.filter(el => el.split("/").length < 2).forEach(el => {
+                    if (this.listRemoteBranches.includes(el)) {
+                        this.util.execSync(`git pull origin ${el}`);
+                    }
+                    this.util.execSync(`git push origin ${el}:${el}`);
+                });
+                resolve();
+            }, 100);
+        })).then(() => {
+            this._onDidChangeTreeData.fire();
+        });
+    }
+    async checkoutBranch(node: Flow | undefined) {
+        let name = node?.full;
+        if (name === undefined) {
+            name = await vscode.window.showQuickPick(
+                this.listBranches.filter(el => el.split("/").length < 2),
+                { title: "Select branch" }
+            );
+        }
+
+        let cmd = `git checkout -q ${name}`;
+
+        this.util.exec(cmd, false, s => {
             this._onDidChangeTreeData.fire();
         });
     }