+++ src/extension.ts
@@ -19,6 +19,12 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('gitflow.init', () => {
         viewBranches.init();
     }));
+    context.subscriptions.push(vscode.commands.registerCommand('gitflow.checkoutBranch', (node?: Flow) => {
+        viewBranches.checkoutBranch(node);
+    }));
+    context.subscriptions.push(vscode.commands.registerCommand('gitflow.syncAll', () => {
+        viewBranches.syncAll();
+    }));
 
 
     context.subscriptions.push(vscode.commands.registerCommand('gitflow.newHotfix', () => {
@@ -119,14 +125,6 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('gitflow.fetchAllBranches', () => {
         viewBranches.fetchAllBranches();
     }));
-
-    context.subscriptions.push(
-        vscode.commands.registerCommand('gitflow.newRelese', () => {
-            viewBranches.startRelease();
-        })
-    );
-
-
 }
 
 export function deactivate() { }