+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "gitflow" extension will be documented in this file.
 
+## [0.2.9] 01/28/22
+  - add - single command to sync all root branches (develop and master ot main)
+  - add - command to checkout root branches
+  - enhance - better icons
+
 ## [0.2.5] - 01/27/22
 
 - fix - Git authenticated commands to remote